Fix Aspire.Hosting.Tests build on VS
When building only the `Aspire.Hosting.Tests` in VS, it would fail with
errors about `ActiveIssue` not being found. This is because the project
was being compiled with none of the dependencies from
`Aspire.Components.Common.Tests` getting used in the
`Aspire.Hosting.Tests` build.

Building the solution worked fine though.

The problem seems to be a duplicate project reference to
`Aspire.Components.Common.Tests`:

```
<ProjectReference Include="..\Aspire.Components.Common.Tests\Aspire.Components.Common.Tests.csproj" IsAspireProjectResource="false" />

<ProjectReference Include="..\Aspire.Components.Common.Tests\Aspire.Components.Common.Tests.csproj" />
```

Removing the second reference fixes the issue. I'm not sure *why* this
broke VS though.