[tests] Follow the same pattern in Hosting tests to create bindMountPath
Instead of `var bindMountPath = Path.Combine(Path.GetTempPath(),
Path.GetRandomFileName());`, and then creating that directory, we can
use `var bindMountPath = Directory.CreateTempSubdirectory().FullName;`
which creates the directory itself, and ensures that it is a new one.

This is already being followed in some other hosting tests.

Also, this ensures that the `bindMountPath` is always created before the
container is started.