Extract Aspire.Hosting.NodeJs.Tests project
Extract the main NodeJs tests from Aspire.Hosting.Tests and the end-to-end TestProject.

Aspire.Hosting.Tests still uses NodeJs in a few places as an example "Executable" resource. Since NodeJs is just being used as an example - and not testing NodeJs itself, I decided to leave Aspire.Hosting.Tests using it.

Also fix an issue with the RequiresTools attribute when the executable isn't an ".exe" extension - like npm is on Windows - it is `npm.cmd`.

Contributes to #4294