Fix configuration binding (#5173)
* Fix configuration binding

The dashboard supports the use of client certificates for authorization to the resource service. This requires a custom resource server, and we're not aware of any that use this feature.

Throughout the Aspire documentation, we advertise values like `Dashboard:ResourceServiceClient:ClientCertificates:Source` to configure certificates, however these values will not bind to the documented configuration values. The CLR property was `ClientCertificates` (plural), not the documented `ClientCertificate` (singular).

This change makes the dashboard respect the documented configuration names.

* Create constants for certificate config names

* Add startup integration test to cover certificate config