[tests] Move more tests to helix by overriding dcp path from assembly metadata (#5215)
`AppHost` build embeds metadata with paths for `DcpCliPath`, `DcpExtensionsPath`, and `DcpBinPath`. This ties the apphost to the machine that it was built on.
This PR:
- When `DcpPublisher:CliPath` is available from the configuration, set the other two paths based on that. 
- Makes `$(RunTestsOnHelix)=true` the default.

This allows running most of the tests on helix. The remaining ones depend on being able to run the test projects which doesn't work on helix because the path to the project files is embedded in the assembly metadata.

Also, download helix test result `.trx` files to have all the `.trx` files available in `artifacts`.

Contributes to https://github.com/dotnet/aspire/issues/2036 .

Co-authored-by: Eric Erhardt <eric.erhardt@microsoft.com>