Hosting.KeyCloak: Remove incorrect param check on `WithRealmImport`
The import path is expected to be relative to the AppHost path, but here
it was being used as-is - relative to the current working directory.
Instead, remove the check and let the downstream `WithBindMount` method
use the path correctly.

This manifested as a failure to import the file when running the app
with `Aspire.Hosting.Testing` which would start the apphost with
cwd!=apphost .