Allow resource properties to be marked as secret
This change allows selected resource properties in the details view to support masking, and show as masked by default. This behaviour exists today for all environment variables, and has so-far not been available for resource properties.

The resource service specifies which properties may be "sensitive", and the UI shows them masked.

As part of this change, we extract a new type, `IPropertyGridItem` that view models must implement to be bound to the `PropertyGrid` component. This allows the component to work directly with the items, rather than calling back to the component's user via parameters for callback logic that largely do the same thing.

We also avoid having the `ResourceDetails` code create copies of `ResourcePropertyViewModel` objects for binding the UI. This copying won't work with the new `IPropertyGridItem` as the `PropertyGrid` requires these to be the same mutable object over time. Removing this fixed the inability to unmask resource property values.