Mask button doesn't jump on mouse over
In details views, we show sensitive values as masked by default. The user has to click a button to show them on screen. This prevents accidentally revealing secrets to onlookers.

The mask button is visible for any sensitive value.

There is also a menu button, that opens a popup to allow copying and visualizing the value. This button is only displayed when the mouse hovers over the row in the name/value grid.

Previously, these were (in order):

- Mask button (always visible)
- Menu button (visible on hover)

Because these were right-aligned, the mask button would be pushed left to make space for the menu button on house hover.

To the user who wanted to toggle a row's mask status, the button they want to press would jump to the left just as the mouse got close. This is surprising and a bit annoying.

This change reorders the buttons so that nothing moves on screen when the mouse hovers over the row.