Show health-check data in dashboard
This change plumbs through data about health-checks to the dashboard.

Values are displayed in the resource detail view, in their own accordian section, in a new three-column grid.

The icon and popup used in the structured log message column to show exceptions was pulled into a new `<ExceptionDetails>` component and reused for health check error data.

Tooltip handling on the resource grid's state column cells was previously inconsistent. We would sometimes add a tooltip to the icon and text, and other times to the cell's entire bounds. Formatting logic for the resource grid's state column is moved into code-behind, simplifying the razor and ensuring consistent treatment. Tooltips are now always at the cell level, making them easier to hit with the mouse, and making the accessible text consistent.

As part of this change, a new `ContentBeforeValue` parameter was added to `GridValue`, as a complement to `ContentAfterValue`.

A new "HealthChecksSandbox" playground was added to make testing of health check data easier.