Put all database connection strings in keyvault (#6059)
* Put all database connection strings in keyvault

When using password authentication and child databases in Azure, don't create a connection string based on the parent's connection string (which is a secretOutputs). The reason being that the deployment tool needs to grab the parent server's connection string from the keyvault and add to it.

Instead, put the database connection strings into the keyvault as well. This way all the secrets are contained in the keyvault and they can be referenced the same way as the parent - via keyvault.

Fix #3299

* Update secret name to be a valid KeyVault secret name.

Since KV secret names can't have underscores, but can have hyphens, we use a hyphen. But this means we need to sanitize the KV secret names when using them in a bicep identifier.