Sanitize bicep indentifiers (#6116)
* Sanitize bicep indentifiers

When we create resources in bicep, we need to ensure the identifier name we use is valid. One common complication is that Aspire resources only allow letters, numbers and a hyphen. But bicep doesn't allow hyphen, but instead an underscore.

Add a new convenience API to get a bicep identifier from an IAzureResource.

Fix #6074

* Add a unit test

* Fix WebPubSub test

* Fix search code to use normalized bicep identifiers