Fix incorrect resource lookups (#6348)
When using IStringLocalizer, we must use the name of the resoure we want to look up. The easiest way to do this is using nameof with the generated property for that resource.

Omitting the namespace causes the resource to be looked up and then used as a key for a lookup, which is not correct. The lookup should happen through IStringLocalizer, not the generated resource class.