Show exception details in text visualizer dialog
Exception details are shown behind a button in two places:

1. In the "Message" column of the structured logs table, when an exception exists.
2. In the "Health reports" grid in resource details, when a health report includes exception details.

Previously, the stack trace would be shown in a popup when the mouse hovered over the icon, sort of like a tooltip.

This commit changes the behaviour so that exception details are shown in the existing text visualizer dialog whenever the icon is clicked.

The icon becomes a `FluentButton` for consistency with the menu button to the right of the cell.

Also some optimisations in `GridValue`:

- Don't use `FluentHighlighter` when the text to highlight is empty, which will be the common case. This improves UI performance.
- Remove `MaxDisplayLength`. We never set this property, and it used space on every `GridValue` instance, and we create a lot of these objects.
- Remove some redundant DOM elements.
- Support adding arbitrary buttons to the right-hand side of the area (used for the "exception details" button).