Add SchedulerHostAddress to DaprSidecarOptions
I enhanced DaprSidecarOptions by adding the SchedulerHostAddress
property. This will allow the developer to specify the host address of
the Dapr scheduler service if not running on localhost (for example, if
developing using Docker Compose, the scheduler service will be running
on a different host).

I enhanced the DaprDistributedApplicationLifecycleHook to pass the
--scheduler-host-address parameter to the Dapr sidecar if the
DaprSidecarOptions.SchedulerHostAddress property is specified.

Resolves dotnet/aspire#6477