Fix Azure Provisioning
With Refactor Bicep parameter usage (dotnet/aspire#6683), we changed when "well-known" parameters were added to the AzureBicepResource. Instead of adding them up front in 2 places, we add them during bicep generation (i.e. ConfigureInfrastructure) and then sync the bicep parameters back to Aspire's AzureBicepResource.Parameters.

This broke Azure Provisioning because it is looking for well known parameters before GetBicepTemplateFile is called. It needs to populate the well-known parameters after.

The fix is to move the calls around so PopulateWellKnownParameters is called after GetBicepTemplateFile.