Allow `WithImage("redis:latest")` and  `AddContainer("redis:latest")`  to work as expected
As a user, I'd expect to be able to take any image name I pass to `docker run` and pass it to `.WithImage(...)` or `.AddContainer(...)` and have it work.
This does not happen today as aspire assumes that you pass the image name and tag as separate components.

This Pr updates WithImage to parse out the tag (or digets) from the image reference, and use that instead of forcing "latest".