Fix AddParameter usage in tests (#7305)
When adding the parameter first, and then setting the config, it accidently works because an exception is thrown at first and then the value is retrieved later when the config is set.

Fix this by ensuring the configuration is set first, before calling AddParameter.

For hosting tests for hosting integrations that aren't part of the core/base APIs, I changed the tests to be simple and just pass the constant value into AddParameter.