Allow CosmosDB connection strings to work with Database and Container keys
With #7408, Cosmos Database and Containers are Aspire Resources now, which means they can be referenced via WithReference.

In the future, the Cosmos Host integration will append these values to the connection string. This will allow the database and container names to be specified outside of the application. (Today they are hard-coded in both, or specified separately in config.) We can also add new client APIs to register a Database or Container object in DI.

This is a stop-gap to support these values in .NET Aspire 9.1 client integration so it doesn't break applications when we start adding these values in the future.

Contributes to #7407