Added WaitForBehavior (#7397)
* Added WaitForBehavior
- This behavior determines what happens to the wait when dependencies are in a failure mode.
The default behavior is determined by if the dashboard is enabled. If enabled the default behavior
is to treat all failures are non terminal. If the dashboard is not enabled the default behavior is to treat
all failures as terminal. This can be overridden by the on a per wait basis.
- Added tests to verify the behavior