Allow EventHub connection strings to work with EntityPath and ConsumerGroup keys
With #7438, EventHub Hub and Consumer Groups are Aspire Resources now, which means they can be referenced via WithReference.

In the future, the EventHubs Host integration will append these values to the connection string. This will allow the hub and consumer group names to be specified outside of the application. (Today they are hard-coded in both, or specified separately in config.) When they flow through the connection string, the EventHubName and ConsumerGroup properties are populated on the settings object.

Contributes to #7407