Allow ServiceBus connection strings to work with EntityPath key (#7523)
In the future, the ServiceBus Host integration will append this value to the connection string. This will allow the queue, topic, and subscription names to be specified outside of the application. (Today they are hard-coded in both, or specified separately in config.) We can also add new client APIs to register a Sender or Processor in DI.

Note that the Subscription can be passed in the EntityPath using the format {topic}/Subscriptions/{subscription}.

This is a stop-gap to support these values in .NET Aspire 9.1 client integration so it doesn't break applications when we start adding these values in the future.

Contributes to #7407