Allow for nested children in the dashboard
Originally the parent-child lookup contained the "root" parent and all the descendants in a flat list. This doesn't show up well in the dashboard because grandchildren are parented directly under their root.

Fix this by making the parent-child lookup only contain direct children, and recursively parent the descendants on their direct parent.

Fix #7580