Support Existing Azure AppInsights, OpenAI, and CosmosDB
These resources don't support existing resources yet. Adding support to finish all the Azure resources for existing.

Note that for CosmosDB, we only explicitly set DisableLocalAuth if WithAccessKeyAuthentication is called. We don't want to disable local auth on an existing resource if it is already enabled as that might break other uses of the resource.

Fix #7665