Support CosmosDB and Redis role assignments on app specific managed identity
We currently don't support an easy API for modifying the role assignments on database resources. There are no role types that can be used from the Azure.Provisioning libraries today. So for now we will just move our existing role assignments to the ACA roles bicep module.

To support database resources, we add a new annotation "RoleAssignmentCustomizationAnnotation" which contains a callback that the resource can implement to inject the role assignment objects into the Azure bicep infrastructure.

Contributes to #6161