Support CosmosDB and Redis role assignments on app specific managed identity (#8140)
* Support CosmosDB and Redis role assignments on app specific managed identity

We currently don't support an easy API for modifying the role assignments on database resources. There are no role types that can be used from the Azure.Provisioning libraries today. So for now we will just move our existing role assignments to the ACA roles bicep module.

To support database resources, we add a new virtual method on AzureProvisioningResource - AddRoleAssignments passing in a context object containing the information on the role assignments to add.

Also needed to split all role assignments out of the {app}-roles bicep module because of an issue with CosmosDB failing if the role assignment is added in the same bicep file as the managed identity it is pointing to. This simplifies the code because the role assignment generation is the same whether the resource is existing or not.

Contributes to #6161
Fix #8142

Also fix tests for new behavior

* - Fix an issue with the AzureResourcePreparer, it should copy default role assignments to applied role assignments even when empty. Add test to ensure.
- Update playground manifests
- Fix tests

* Add more tests

* Split all role assignments into separate bicep modules.

This works around an issue with CosmosDB where creating a managed identity and assigning a CosmosDB role assignment in the same bicep module fails.

* Refactor how role assignments are created

Add a new virtual method on AzureProvisioningResource to add role assignments. This method is overriden in database resources (like CosmosDB and Redis) to customize how the role assignments are created.

* Clean up

* Fix tests

* PR feedback

fill out XML docs

* PR feedback