Fix Azure Functions Role Assignments to Host Storage (#8290)
* Fix Azure Functions Role Assignments to Host Storage

With #8127, a mistake was made in that WithRoleAssignments only
works if the app is using ACA infrastructure. If the app isn't using our
ACA infrastructure, WithRoleAssignments throws an exception at startup.

To make Azure Functions work correctly both with and without ACA infrastructure
we use the internal WithDefaultRoleAssignments method to set the default
role assignments on the Host Storage, which will be respected for both.

* Add FunctionsTest without AzureContainerApps