Run only quarantined tests in outer loop
Allow quarantine flaky tests, which excludes the test from being executed
by the normal CI pipelines. The quarantined tests get executed on preset
schedule by a cron job for some period of time, after which they could
manually unquarantined, if their stability improves.

The quarantine attribute and the discoverer as well as some of the MSBuild
infra are copied from dotnet/aspnetcore.

The test execution is implemented in an Arcade SDK way, which should
make it easy to maintain and enhance, if required.

The quarantined test can be invoked with the following command:

    ./build.cmd -projects ./tests/Shared/SolutionTests.proj
                -restore -build -test
                /p:IsGitHubActions=true
                /p:RunQuarantinedTests=true

Setting "RunQuarantinedTests=false" (which is default, so can be omitted)
will execute all the non-quarantined tests.