Use a new $(SkipTestProjects) property to completely skip test projects.
build scripts's `-testnobuild` is relevant only when `-test` is being
used, and it avoids invoking build on the test projects before running
them.

But `$(SkipTestProjects)` allows skipping test projects from being
included in the build at all. This is useful for cases like when we are
just building the packages, and don't need to build the test projects.