[CI] Fix pipeline tests on rolling builds (#8333)
* [CI] Fix handling `testnobuild`

Use `$(VSTestNoBuild)` property which is used by vstest instead of
arbitrary `$(TestNoBuild)`. VSTest prevents a build of the test projects
when the property is `true`.

This fixes the running tests on Azdo pipeline builds.

Fixes https://github.com/dotnet/aspire/issues/8330 .

* Address review feedback from @ russkie

* Use a new $(SkipTestProjects) property to completely skip test projects.

build scripts's `-testnobuild` is relevant only when `-test` is being
used, and it avoids invoking build on the test projects before running
them.

But `$(SkipTestProjects)` allows skipping test projects from being
included in the build at all. This is useful for cases like when we are
just building the packages, and don't need to build the test projects.