Fix Azure SqlServer deployment with ACA environment
Azure SqlServer needs an admin when it is first provisioned. To supply this, we use the principalId from the Azure Container App Environment. But today the ACA environment is filling it with the wrong ID - it is using the managedIdentity.Id and not PrincipalId.

To fix this, fill in the empty principalId parameters with the correct principalId

Contributes to #8381