[tests] Simplify some dashboard part of template tests (#8475)
Reading the `Endpoints` from the dashboard for various resources is very
fragile as it depends on html layout, which can change, and can have
overflows etc.

Instead, we drop checking for Endpoints from the dashboard, as it should
be getting validated in other ways. And we use `AspireProject.InfoTable`
to get the endpoints, and check them.

Fixes https://github.com/dotnet/aspire/issues/8473 .