AddAzureContainerAppEnvironment should use the environment name as a prefix (#8606)
* AddAzureContainerAppEnvironment should use the environment name as a prefix

We should be prefixing the resources created via AddAzureContainerAppEnvironment with the environment name. This will allow for multiple environments to be in a single distributed application in the future.

With this change, it means deploying to existing environments will duplicate resources. To solve that, add a new method WithAzdResourceNaming, which will revert the resource names back to the previous naming scheme.

* Revert local testing changes

* Ensure resourceToken is declared before it is used in bicep

* Support volume names