Add IResourceWithParameters and use it in the Azure Publisher (#9064)
* Add IResourceWithParameters and use it in the Azure Publisher

There are external resources that can reference an AzureBicepResource's output. Today the Azure Pubilsher only looks at and "lifts" outputs that are referenced by DeploymentTargetAnnotations on resources. But there are other kinds of resources that can have parameters referencing Azure bicep outputs.

Along with this, the AzurePublishingContext should expose 2 maps:

- A map of Aspire ParameterResources to the main.bicep's ProvisioningParameter that was generated to take the parameter value.
- A map of Aspire BicepOutputReferences to the main.bicep's ProvisioningOutput that was lifted from a bicep resource.

Also, implement IEquatable on BicepReferenceOutput, so they can be used as keys in a dictionary.

* Rename Map and use concrete Dictionary.

* Update src/Aspire.Hosting.Azure/AzureBicepResource.cs

---------

Co-authored-by: David Fowler <davidfowl@gmail.com>