Set functionapp kind on Azure Functions projects (#9779)
* Set Azure Functions kind on ACA containers

Introduce a new interface `IAzureFunctionsResource` to represent Azure Functions resources in Aspire.
This interface allows for the configuration of Azure Functions resources, including setting the kind of Azure Functions to be used with Azure Container Apps (ACA).

This change ensures that when Azure Functions are deployed as part of an Azure Container App, the kind is set to "functionapp" by default.
This is necessary to ensure that the Azure Functions are correctly recognized and managed within the Azure Container Apps environment.

Fix #8188

* Use an annotation instead of an interface.