Reenable extensions interaction service (#10047)
* Reapply "Initial extension interaction service (#9927)" (#10030)

This reverts commit aab8668e48485f603341d269b21b7d77c53e3efb.

* Update ExtensionBackchannel to be AOT compatible

Reuse the same Json source generation context. Add ExtensionBackchannel types to it.

Need to suppress a RDC warning due to EventHandlers on the Rpc Target object needing to call MakeGenericMethod, but we aren't using EventHandlers so it can be suppressed.

* Fix new AOT warnings:

Remove Exception from the wire payload since it can't be safely STJ serialized using the source generator.

Enable Configuration Binder source generator in the Aspire.Cli.

* untrack generated files

* subclass OperationCanceledException for ExtensionInteractionService

---------

Co-authored-by: Adam Ratzman <adam@adamratzman.com>