Brackets in ReferenceExpressions should be escaped in the manifest
The aspire manifest uses `{foo}` for references to other resources in the manifest. But when a value wants an explicit bracket in the ValueExpression, it needs to be escaped in the manifest.

This change ensures that hard-coded brackets in ReferenceExpressions are turned into escaped double brackets in the manifest/ValueExpression.

Fix #5525