Aspire cli: Add AOT builds for Linux, and macOS (#10275)
Add AOT builds for Linux, and macOS

## What?

This PR adds AOT builds for Linux, and macOS on the internal builds. On macOS the binaries are appropriately signed, and notarized making them usable after downloading.

## How?

This is done by having separate OS-specific jobs just to build the AOTed aspire cli. And these upload the archives for use by a "joining" Windows job which builds windows binaries, and publishes all the archives. Checksum (`sha512`) files are also generated for the archives which can be used to verify the archive when downloading.

This is the current supported matrix:

| Runtime Identifier | AOTed |
|-------------------|-------------|
| `win-x64` | ✅ |
| `win-arm64` | ✅ |
| `win-x86` | ❌ |
| `linux-x64` | ✅ |
| `linux-arm64` | ❌ |
| `linux-musl-x64` | ❌ |
| `osx-x64` | ✅ |
| `osx-arm64` | ✅ |

- `win-x86` - Native AOT is not supported for this
- `linux-arm64`, `linux-musl-x64` - will be added in a follow up PR

For the non-aot cases a self-contained executable is generated.

## Tests

To catch missing archives due to any reason on the CI build

--
* Addresses review feedback from @ eerhardt
* Addresses review feedback from @ joperezr

Co-authored-by: Eric Erhardt <eric.erhardt@microsoft.com>