Use CreateEmptyApplicationBuilder in Aspire.Cli
We are currently creating a default HostApplicationBuilder, which tries to load appsettings.json files in the CWD. This in turn is creating a FileWatcher for the CWD and all directories under it. Causing the startup of the CLI to suffer depending on the contents of the CWD.

The fix is to use an empty HostApplicationBuilder, which doesn't add appsettings.json config files by default. It also means we don't have to clear logging providers anymore, because there aren't any in an empty builder.

Fix #9765