aspire-cli: Fix linux-arm64 and linux-musl-x64 to produce tarballs and not zip files (#10545)
* aspire-cli: always use the correct ArchiveFormat

Problem:

We produce `linux-arm64` archives on a Windows azdo agent. And the
Archives targets defaults to `.zip` Windows. Even though we set that
property in `Common.projitems` it does not take effect as it is imported
after the props files.

Solution:

Always override it to the intended values.

* Build linux-arm64, and linux-musl-x64 on individual build agents

- These rids were being built on Windows because we could not Native AOT them as
only the x64 linux agents were being used.
- Creating tarballs for these fails on windows - https://github.com/dotnet/arcade/issues/15982
- Instead build these on linux agents so we can generate the tarballs
  too.

* Add a test to ensure the correct cli archives files are generated

Generating tarballs on linux wasn't working but it did not fail the
build because - https://github.com/dotnet/arcade/issues/15982 .

Add a msbuild based check that confirms the cli archives have the
expected extensions:

- Windows - .zip
- non-windows - .tar.gz