Allow requiring OpenTelemetry protocol if needed (#10507)
* Allow requiring OpenTelemetry protocol if needed

For some scenarios (mine is runnin it ASP.NET framework applications), gRPC is not supported. This change adds an overload for AddOtlpExporter APIs to have a required protocol. Since there's not a default http endpoint if one is not set, this will throw an exception pointing the user to set that variable.

* add comment and move private method down

* react to comment changes from merge