Rename Aspire.Hosting.NodeJs to Aspire.Hosting.JavaScript
This makes the library consistent to use language naming, and not a single technology. It allows for more JS features in the future that aren't tied to Node.js.

Since the library is being renamed, I also removed the Obsolete APIs since there is no reason to ship them in the new library.

Contributes to #12058