[tests] Add Snapshots directory to the tests archive (#13323)
.. and on helix use from there.

The helix targets were copying files from all the `Snapshots` directories in the various test projects into a single directory, and then using that for all the test runs. But some of them would overwrite each other because of the same name causing tests to fail.

Instead, we copy the correct `Snapshots` directory to the tests archive generated for helix. And the test run on helix uses them from the subdirectory.

Fixes https://github.com/dotnet/aspire/issues/13298