Updated versions for Azure.Provisioning and Azure.Provisioning.AppService packages (#13281)
Update to the latest stable versions of Azure.Provisioning and Azure.Provisioning.AppService. This includes updating Azure.Core as well, since it is a transitive dependency.

With this change, we are changing all ProvisioningOutput.Value to use the new `ToBicepExpression()` function which ensures the bicep is written like a human would write it. For example, it uses the resource's property as an expression instead of duplicating the values in multiple places:

```bicep
param existingResourceName string

resource redis 'Microsoft.Cache/redisEnterprise@2025-04-01' existing = {
  name: existingResourceName
}

output name string = existingResourceName // OLD
output name string = redis.name // NEW
```

---------

Co-authored-by: Shilpi Rachna <shilpir@microsoft.com>
Co-authored-by: Eric Erhardt <eric.erhardt@microsoft.com>