Point NPM_CONFIG_USERCONFIG to root .npmrc when installing yarn/vsce globally (#13358)
When using the -g (global) flag for npm install it will ignore the .npmrc in the current working directory. Instead we need to pass it as an environment variable to make sure our custom npm feed is used.

Also remove the always-auth entry since it is deprecated/unused in npm and will be removed: https://github.com/actions/setup-node/issues/1305

This fixes the registry.npmjs.org related violations for SFI network isolation, it needs to be backported to all active branches.