feat: add custom Orleans provider type annotation
The algorithm of deriving the provider type from the resource type name does not work in every case.
For example, if you want to use Garnet as a drop-in replacement for Redis,
Orleans does not find the corrent provider builder, because there is none defined for "Garnet".

With the new `OrleansProviderTypeAnnotation`, users can explicitly specify the provider type to use for a given resource.

Example:

```
var garnet = builder.AddGarnet("garnet")
                    .WithOrleansProviderType("Redis");

var orleans = builder.AddOrleans()
                     .WithClustering(garnet);
```